refactor: ファイル名をbirthdayからbirthdaySimpleへリネームし、エクスポートを更新
