+++ package/main/src/Simple/Tool/index.ts
@@ -1,3 +1,3 @@
-export * from "./birthday";
+export * from "./birthdaySimple";
 export * from "./dayOfWeekSimple";
 export * from "./deviationValueSimple";
